Dynamic Levels
Created M_PersistentWorld, which contains a block out for a new tutorial level, and a proof of concept for the Dynamic Levels system. Levels tend to spawn on top of each other, however, which results in unintended effects. Possible solutions are smarter level spawning, and making specific gameplay-related actors intelligently despawn when clipping.